+++ alarm/src/main/java/fr/radiofrance/alarm/receiver/RfAlarmReceiver.java
@@ -12,8 +12,10 @@ import fr.radiofrance.alarm.manager.RfAlarmManager;
 import fr.radiofrance.alarm.notification.AlarmNotificationManager;
 import fr.radiofrance.alarm.util.AlarmIntentUtils;
 
-
 public class RfAlarmReceiver extends BroadcastReceiver {
+
+    public static final String ACTION_BROADCAST_RECEIVER_ON_ALARM_NEED_UI_REFRESH = "fr.radiofrance.alarm.receiver.ACTION_BROADCAST_RECEIVER_ON_ALARM_NEED_UI_REFRESH";
+    public static final String EXTRA_ALARM_ID_KEY = "fr.radiofrance.alarm.receiver.EXTRA_ALARM_ID_KEY";
 
     public static void enable(final Context context) {
         setReceiverEnabledSetting(context, true);
@@ -67,10 +69,17 @@ public class RfAlarmReceiver extends BroadcastReceiver {
 
     private void onAlarmNotificationCancel(final Context context, final Intent intent) {
         try {
-            RfAlarmManager.with(context)
-                    .onAlarmNotificationCancelBroadcastReceived(intent.getStringExtra(AlarmNotificationManager.EXTRA_ALARM_NOTIFICATION_ALARM_ID_KEY),
-                            intent.getLongExtra(AlarmNotificationManager.EXTRA_ALARM_NOTIFICATION_TIME_MILLIS_KEY, -1L),
-                            intent.getBooleanExtra(AlarmNotificationManager.EXTRA_ALARM_NOTIFICATION_IS_SNOOZE_KEY, false));
+            final String alarmId = intent.getStringExtra(AlarmNotificationManager.EXTRA_ALARM_NOTIFICATION_ALARM_ID_KEY);
+            final boolean isSnooze = intent.getBooleanExtra(AlarmNotificationManager.EXTRA_ALARM_NOTIFICATION_IS_SNOOZE_KEY, false);
+
+            RfAlarmManager.with(context).onAlarmNotificationCancelBroadcastReceived(
+                    alarmId,
+                    intent.getLongExtra(AlarmNotificationManager.EXTRA_ALARM_NOTIFICATION_TIME_MILLIS_KEY, -1L),
+                    isSnooze);
+
+            if (!isSnooze) {
+                context.sendBroadcast(new Intent(ACTION_BROADCAST_RECEIVER_ON_ALARM_NEED_UI_REFRESH).putExtra(EXTRA_ALARM_ID_KEY, alarmId));
+            }
         } catch (RfAlarmException e) {
             e.printStackTrace();
         }